requirements.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
dj-database-url==0.5.0
Django==4.0.2
django-dotenv==1.4.2
django-filter==21.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.6.1
pipreqs==0.4.11
platformdirs==2.5.0
psycopg2-binary==2.9.3
pylint==2.12.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==6.0.0
wrapt==1.13.3
yarg==0.1.9
zipp==3.7.0
